requirements.txt
numpy
pandas
seaborn
matplotlib
mysql-connector-python
pymysql
python-dotenv
scikit-learn